Camera controller, modified player controller
Added a new script that makes the camera follow the player, changed the
view to be near 90 degree top-down, added a new folder for textures
along with a grid texture. Changed the floor material to the grid and
made the player red.
In the PlayerController script I changed the Input.GetAxis("...")
methods to Input.GetAxisRaw("...").
This fixed the slidey problem we had